requirements.txt
aiogram==3.13.1
httpx==0.27.0
feedparser==6.0.11
APScheduler==3.10.4
Pillow==10.4.0
python-dotenv==1.0.1
